Major updates
Updated to slightly following the execution of ulexec as opposed to the
python version. The stack was the primary issue with the previous
version. Now we have a version that should work with Go binaries (Go
really needs the stack and auxv setup correctly...)

Also, removed ureq by default and placed behind a feature flag to reduce
binary size when web requests aren't needed